fix: Resolve nephio-generator type mismatches in test interfaces
Fixed compilation errors in nephio_integration_test.go by creating separate mock implementations for generator and renderer packages:
- Created MockGeneratorFunctionRegistry implementing generator.KptFunctionRegistry
- Created MockRendererFunctionRegistry implementing renderer.FunctionRegistry
- Updated MockPackageValidator to implement correct interface methods
- Updated MockRenderValidator to return proper ValidationResult structs
- Added comprehensive unit tests to verify mock implementations

The issue was that the same mock was trying to implement two different interfaces with incompatible KptFunction types.

+++ tn/tests/unit/tc_test.go
@@ -329,7 +329,7 @@ func TestIperfManager(t *testing.T) {
 				if tc.valid {
 					assert.NotEmpty(t, tc.config.ServerIP)
 					assert.Greater(t, tc.config.Port, 0)
-					assert.Greater(t, tc.config.Duration, 0)
+					assert.Greater(t, tc.config.Duration, int64(0))
 					assert.Contains(t, []string{"tcp", "udp"}, tc.config.Protocol)
 				}
 			})
